refactor(agentic): remove operating-model personalization layer (product logic → semteams)
The operating-model / "personalization" layer — per-user operating-model
profiles built via an /onboard interview, compaction "lessons", and
profile-context injection — was upstreamed into the framework prematurely
(c2a10d4b, ADR-025 Phase 2). It is PRODUCT logic, not framework substrate:

  - Every entity it writes is `{org}.{platform}.user.teams.*` (a teams-product
    namespace); the producer subject is `teams.operating_model.layer_approved.*`.
  - None of it is a primitive another product composes — it's one product's
    domain model (five canonical layers, the onboard interview, the profile
    preamble).
  - It is wired in NO semstreams config and its consumer wiring never worked
    (setupInputConsumers diverged from handlerForPort; no config declared the
    ports). It was dead in the framework.

semstreams owns the substrates this stood on (graph-ingest mutation lanes, the
ADR-054 indexing-profile substrate, the payload registry, the rule engine);
those stay. The product feature belongs in semteams, which can redesign and
re-consume the substrates if it wants the feature back.

Removed:
  - agentic/operating-model/ (entire package)
  - agentic-memory: layer_approved + profile_context + profile_support + lesson
    path + port_routing (kept the generic extraction / hydration / checkpointing
    core that deep-research uses)
  - agentic-dispatch: /onboard command + onboarding interview/tracker +
    normalize_extractor (kept core dispatch + intent_classifier)
  - agentic-loop: profile_context_handler + inject_profile_context config
  - payloadbuiltins + graph-ingest registry seeds + drift-test cases + schemas

Kept the generic dispatch / loop / memory cores intact.

Validation: build + `go test -race ./...` + integration sweep (dispatch/loop/
memory/graph-ingest) + e2e:agentic + e2e:deep-research all green; schema
regenerated; no config/runtime references the removed features. The graph-ingest
bare-update envelope-preservation footgun this work surfaced is fixed separately
(#258).
