+++ processor/agentic-dispatch/config.go
@@ -18,7 +18,6 @@ type Config struct {
 	DeleteConsumerOnStop       bool                  `json:"delete_consumer_on_stop,omitempty" schema:"type:bool,description:Delete durable consumers on Stop (use for tests only),category:advanced,default:false"`
 	Ports                      *component.PortConfig `json:"ports,omitempty" schema:"type:ports,description:Port configuration for inputs and outputs,category:basic"`
 	EnableIntentClassification bool                  `json:"enable_intent_classification,omitempty" schema:"type:bool,description:Enable LLM-assisted intent classification for ambiguous messages,category:advanced,default:false"`
-	EnableOnboarding           bool                  `json:"enable_onboarding,omitempty" schema:"type:bool,description:Enable /onboard command for operating model interview,category:advanced,default:false"`
 	// DefaultTools is the tool allowlist for initial-user-message tasks
 	// (the first loop in a chain, before any rule has fired). Names are
 	// resolved against the agentictools global registry at dispatch time;
