+++ processor/agentic-loop/config.go
@@ -178,11 +178,10 @@ type ContextConfig struct {
 	HeadroomRatio    float64 `json:"headroom_ratio" description:"Fraction of model context to reserve for responses (0.0-0.5). Takes precedence over headroom_tokens when the computed value is larger"`
 	HeadroomTokens   int     `json:"headroom_tokens" description:"Minimum token headroom floor — ratio-based headroom never goes below this value"`
 	// Multi-agent context budget fields
-	MaxBudgetTokens      int      `json:"max_budget_tokens,omitempty" description:"Hard token limit for context budget (overrides model limits when set)"`
-	SliceOnBudget        bool     `json:"slice_on_budget,omitempty" description:"Enable context slicing when budget is exceeded"`
-	PreserveEntities     []string `json:"preserve_entities,omitempty" description:"Entity IDs to always keep in context during slicing"`
-	EntityPriority       int      `json:"entity_priority,omitempty" description:"Priority for entity context vs conversation (1-10, higher = more entity context)"`
-	InjectProfileContext bool     `json:"inject_profile_context,omitempty" description:"Inject operating model profile context into system prompt when available"`
+	MaxBudgetTokens  int      `json:"max_budget_tokens,omitempty" description:"Hard token limit for context budget (overrides model limits when set)"`
+	SliceOnBudget    bool     `json:"slice_on_budget,omitempty" description:"Enable context slicing when budget is exceeded"`
+	PreserveEntities []string `json:"preserve_entities,omitempty" description:"Entity IDs to always keep in context during slicing"`
+	EntityPriority   int      `json:"entity_priority,omitempty" description:"Priority for entity context vs conversation (1-10, higher = more entity context)"`
 }
 
 // Validate validates the configuration
