+++ processor/agentic-memory/component.go
@@ -13,7 +13,6 @@ import (
 	"sync/atomic"
 	"time"
 
-	operatingmodel "github.com/c360studio/semstreams/agentic/operating-model"
 	"github.com/c360studio/semstreams/component"
 	"github.com/c360studio/semstreams/message"
 	"github.com/c360studio/semstreams/natsclient"
@@ -33,9 +32,8 @@ type Component struct {
 	platform   component.PlatformMeta
 	decoder    *message.Decoder
 
-	hydrator      *Hydrator
-	extractor     *LLMExtractor
-	profileReader atomic.Pointer[operatingmodel.ProfileReader]
+	hydrator  *Hydrator
+	extractor *LLMExtractor
 
 	// Lifecycle management
 	running   bool
@@ -91,7 +89,6 @@ func NewComponent(rawConfig json.RawMessage, deps component.Dependencies) (compo
 		hydrator:   hydrator,
 		extractor:  extractor,
 	}
-	initProfileReader(&comp.profileReader)
 	return comp, nil
 }
