+++ processor/agentic-memory/handlers.go
@@ -7,9 +7,6 @@ import (
 	"time"
 
 	"github.com/c360studio/semstreams/agentic"
-	operatingmodel "github.com/c360studio/semstreams/agentic/operating-model"
-	"github.com/c360studio/semstreams/message"
-	"github.com/google/uuid"
 )
 
 // ContextEvent is the compaction event shape this handler decodes. It is a
@@ -111,63 +108,6 @@ func (c *Component) handleCompactionComplete(ctx context.Context, event ContextE
 	c.logger.Debug("Post-compaction context injected",
 		"loop_id", event.LoopID,
 		"token_count", hydrated.TokenCount)
-
-	// Persist the compaction summary as a Lesson on the user's profile so
-	// future loops can fold it into ProfileContext.LessonsLearned. Best
-	// effort — a publish failure is logged but does not retry, since the
-	// context has already been hydrated for the current loop and a missed
-	// lesson is a degradation of future-session quality, not a hard error.
-	c.persistCompactionAsLesson(ctx, event)
-}
-
-// persistCompactionAsLesson converts a compaction_complete event's summary
-// into a Lesson entity attached to the owning user's profile. No-op when
-// the event has no UserID (e.g. system-initiated loops, pre-beta.29 events
-// from older publishers), no summary, or when the component's platform
-// metadata is unconfigured — Org/Platform empty would otherwise panic in
-// LessonEntityID/ProfileEntityID, fail safeHandleMessage's recover, and
-// trigger a NATS redeliver loop into DLQ.
-func (c *Component) persistCompactionAsLesson(ctx context.Context, event ContextEvent) {
-	if event.UserID == "" || event.Summary == "" {
-		return
-	}
-	if c.platform.Org == "" || c.platform.Platform == "" {
-		c.logger.DebugContext(ctx, "lesson persist skipped: platform unconfigured",
-			"loop_id", event.LoopID, "user_id", event.UserID)
-		return
-	}
-	lesson := operatingmodel.Lesson{
-		LessonID:  "lesson-" + uuid.New().String()[:8],
-		Summary:   event.Summary,
-		SessionID: event.LoopID,
-		LearnedAt: time.Now().UTC(),
-	}
-	if err := lesson.Validate(); err != nil {
-		c.logger.WarnContext(ctx, "lesson validation failed; skipping persist",
-			"loop_id", event.LoopID, "user_id", event.UserID, "error", err)
-		return
-	}
-
-	triples := operatingmodel.LessonTriples(operatingmodel.ProfileRef{
-		Org:      c.platform.Org,
-		Platform: c.platform.Platform,
-		UserID:   event.UserID,
-	}, lesson)
-
-	if err := c.publishLessonTriples(ctx, event.LoopID, triples); err != nil {
-		c.logger.WarnContext(ctx, "lesson publish failed; future sessions will miss this insight",
-			"loop_id", event.LoopID, "user_id", event.UserID, "error", err)
-		return
-	}
-	c.logger.Debug("compaction summary persisted as lesson",
-		"loop_id", event.LoopID, "user_id", event.UserID, "lesson_id", lesson.LessonID)
-}
-
-// publishLessonTriples sends lesson triples through the same graph_mutations
-// output port the existing fact-extraction path uses. graph-ingest applies
-// them via AddTriple → CAS write to ENTITY_STATES.
-func (c *Component) publishLessonTriples(ctx context.Context, loopID string, triples []message.Triple) error {
-	return c.publishGraphMutations(ctx, loopID, "add_triples", triples)
 }
 
 // handleCompactionStarting processes compaction start events
